NEON #13: retype AcMainTask drag anchor to float (disasm 0x99e3e)
The arcade drag-scroll anchor m_dragAnchorX/Y (@+0x50c/+0x510) is stored by the
binary as a plain float (vcvt.f32.s32 of the touch coord + vstr.32) -- the
reconstruction typed them int and divided the down-latch by 65536, a divide the
binary does not do. The fields are write-only in the reconstruction (their
consumer -- the per-frame scroll normalization delta=((float)touch-anchor)/
m_screenScale in the not-yet-reconstructed arcade states 0x10/0x4d of the 24KB
AcMainTask::update -- is absent), so the retype is a zero-risk fidelity fix.
Residual scroll-norm states documented in NEON_ACCURACY.md #13.

+++ Project/System/src/Task/AcMainTask.h
@@ -168,8 +168,8 @@ private:
     uint8_t          _rsvd_4f2[0x4f4 - 0x4f2] = {};   // +0x4f2
     Random           m_rng;                         // +0x4f4 embedded arcade RNG (ctor placement-constructs; dtor destroys)
     int             m_dragAnchorId = {};            // +0x508 touch drag anchor id (-1 == none)
-    int             m_dragAnchorX = {};             // +0x50c drag anchor x
-    int             m_dragAnchorY = {};             // +0x510 drag anchor y
+    float           m_dragAnchorX = {};             // +0x50c drag anchor x (float; disasm 0x99e3e)
+    float           m_dragAnchorY = {};             // +0x510 drag anchor y (float)
     float           m_scrollBaseX = {};             // +0x514 scroll base x (subtracted for screen space)
     float           m_scrollBaseY = {};             // +0x518 scroll base y
     uint8_t          _rsvd_51c[0x524 - 0x51c] = {};   // +0x51c
